+++ CoalStorage.Core/Interfaces/IBaseRepository.cs
@@ -2,10 +2,10 @@
 
 public interface IBaseRepository<T> where T : class
 {
-    Task<T> GetByIdAsync(int id);
+    Task<T> GetByIdAsync(long id);
     Task<IEnumerable<T>> GetAllAsync();
     Task AddAsync(T entity);
     Task UpdateAsync(T entity);
-    Task DeleteAsync(int id);
+    Task DeleteAsync(long id);
     Task SaveChangesAsync();
 }
